splitted dropoff card
